requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
PyJWT
pydantic
python-jose
passlib[bcrypt]
email-validator
pydantic[email]
loguru
pytest
